Further speedup for fattal, also removed some duplicate code
